That's Odd
Task:
Find the sum of all even integers in a list of numbers.

Input Format:
The first input denotes the length of the list (N). The next N lines contain the list elements as integers.

Output Format:
An integer that represents the sum of only the even numbers in the list.