It's a sign
Task
Given the four words that were supposed to be contained in each box, determine if at least one of them is of a palindrome.

Input Format
Four strings that represent the word or phrase that was supposed to be printed on the signs inside each box.

Output Format
A string that say 'Open' if at least one of the boxes is a palindrome or 'Trash' if all of the signs are misprinted.